updated install script so that it creates a proper vrep/system/usrset.txt, not an empty one. previously an empty one was being created since the vrep package does not come with this file. it is created the first time vrep runs.

+++ install.sh
@@ -182,10 +182,16 @@ function install_vrep {
         echo "Unpacking V-REP ..." | tee -a $LOG
         mkdir vrep
         tar xfz $VREP_PKG -C vrep  --strip-components 1
-        for i in doNotShowOpenglSettingsMessage doNotShowCrashRecoveryMessage doNotShowUpdateCheckMessage; do
-            cat vrep/system/usrset.txt | sed "s/$i = false/$i = true/g" > vrep/system/usrset.txt1
-            mv vrep/system/usrset.txt1 vrep/system/usrset.txt
-        done
+        if [ -f vrep/system/usrset.txt ]; then
+            for i in doNotShowOpenglSettingsMessage doNotShowCrashRecoveryMessage doNotShowUpdateCheckMessage; do
+                cat vrep/system/usrset.txt | sed "s/$i = false/$i = true/g" > vrep/system/usrset.txt1
+                mv vrep/system/usrset.txt1 vrep/system/usrset.txt
+            done
+        else
+            for i in doNotShowOpenglSettingsMessage doNotShowCrashRecoveryMessage doNotShowUpdateCheckMessage; do
+                echo "$i = true" >> vrep/system/usrset.txt
+            done
+        fi
     fi
 }
